Add NoOpMetricProvider to the DefaultMetricRegistry
